+++ src/modules/products/products.service.ts
@@ -1,17 +1,48 @@
-import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  HttpException,
+  HttpStatus,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { Product, ProductDocument } from './entities/product.entity';
 import { InjectModel } from '@nestjs/mongoose';
 import { ProductDto, UpdateProductDto } from './dto/product.dto';
 import { Model } from 'mongoose';
+import { sanitizedName } from 'src/shared/utils/helpers';
+import { BucketService } from 'src/shared/bucket/bucket.service';
 
 @Injectable()
 export class ProductsService {
   constructor(
     @InjectModel(Product.name) private productModel: Model<ProductDocument>,
+    private readonly bucketService: BucketService,
   ) {}
 
-  create(createProductDto: ProductDto) {
-    return this.productModel.create(createProductDto);
+  async create(createProductDto: ProductDto) {
+    const publicId = `products/${sanitizedName(createProductDto.name)}`;
+
+    const existingProduct = await this.productModel.findOne({
+      publicIdImage: publicId,
+    });
+
+    if (existingProduct) {
+      throw new ConflictException(
+        'Já existe um produto com esse nome (imagem).',
+      );
+    }
+
+    const imageUrl = await this.bucketService.uploadBase64Image(
+      createProductDto.urlImage,
+      'products',
+      sanitizedName(createProductDto.name),
+    );
+    const product = new this.productModel({
+      ...createProductDto,
+      urlImage: imageUrl,
+      publicIdImage: publicId,
+    });
+    return product.save();
   }
 
   async findAll() {
@@ -50,7 +81,6 @@ export class ProductsService {
     return product;
   }
 
-  // Atualizar um produto
   async update(id: string, updateProductDto: UpdateProductDto) {
     const updatedProduct = await this.productModel
       .findByIdAndUpdate(id, updateProductDto, {
@@ -67,7 +97,14 @@ export class ProductsService {
     return updatedProduct;
   }
 
-  remove(id: string) {
+  async remove(id: string) {
+    const product = await this.productModel.findById(id);
+
+    if (!product) {
+      throw new NotFoundException('Product not found');
+    }
+
+    await this.bucketService.deleteImageByName(product.publicIdImage);
     return this.productModel.findByIdAndDelete(id);
   }
 }
